requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
assertpy==1.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
curlify==2.2.1
distconfig3==1.0.1
Faker==33.0.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
PyHamcrest==2.1.0
pytest==8.3.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
structlog==24.4.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.3
vyper-config==1.2.1
watchdog==6.0.0
